requirements.txt
paho-mqtt
fastapi[standard]
sqlmodel
psycopg2-binary
python-dotenv
uvicorn
sqlalchemy